feat: Updates to core functionality and dynamic UI updates
Add sass files to entry index.js script.
Updates to app to capture relevant info that will be added to backend server data.
Updates to UI to show details of trip in the correct section.
Add resetUI to reset the trip details back to default and tie it to remove trip button.
Semantic updates to index.html as well as placeholder icons and links for visual appeal.
Updates to server variable names for better consistency.